Allow Manifest to grow as a log for 1000 items before rolling over to speed up the commit time.

+++ RazorDB/Manifest.cs
@@ -19,6 +19,11 @@ namespace RazorDB {
             get { return _baseFileName; }
         }
 
+        private int _manifestVersion = 0;
+        public int ManifestVersion {
+            get { return _manifestVersion; }
+        }
+
         private const int num_versions = 10;
         private int[] _versions = new int[num_versions];
         public int CurrentVersion(int level) {
@@ -45,42 +50,72 @@ namespace RazorDB {
             string manifestFile = Config.ManifestFile(_baseFileName);
             string tempManifestFile = manifestFile + "~";
 
-            // Make a backup of the current manifest file
-            if (File.Exists(manifestFile))
-                File.Move(manifestFile, tempManifestFile);
+            _manifestVersion++;
 
-            FileStream fs = new FileStream(manifestFile, FileMode.CreateNew, FileAccess.Write, FileShare.None, 1024, false);
-            BinaryWriter writer = new BinaryWriter(fs);
+            if (ManifestVersion > Config.ManifestVersionCount) {
+                // Make a backup of the current manifest file
+                if (File.Exists(manifestFile))
+                    File.Move(manifestFile, tempManifestFile);
 
-            try {
-                writer.Write(_versions.Length);
-                foreach (var b in _versions) {
-                    writer.Write(b);
+                FileStream fs = new FileStream(manifestFile, FileMode.CreateNew, FileAccess.Write, FileShare.None, 1024, false);
+                BinaryWriter writer = new BinaryWriter(fs);
+
+                try {
+                    WriteManifestContents(writer);
+                    writer.Close();
+                } catch {
+                    writer.Close();
+                    File.Delete(manifestFile);
+                    File.Move(tempManifestFile, manifestFile);
+                    throw;
                 }
-                writer.Close();
-            } catch {
-                writer.Close();
-                File.Delete(manifestFile);
-                File.Move(tempManifestFile, manifestFile);
-                throw;
+
+                // Delete the backup file
+                if (File.Exists(tempManifestFile))
+                    File.Delete(tempManifestFile);
+            } else {
+                FileStream fs = new FileStream(manifestFile, FileMode.Append, FileAccess.Write, FileShare.None, 1024, false);
+                BinaryWriter writer = new BinaryWriter(fs);
+                try {
+                    WriteManifestContents(writer);
+                } finally {
+                    writer.Close();
+                }
+            }
+        }
+
+        private void WriteManifestContents(BinaryWriter writer) {
+            long startPos = writer.BaseStream.Position;
+
+            writer.Write(ManifestVersion);
+            writer.Write(_versions.Length);
+            foreach (var b in _versions) {
+                writer.Write(b);
             }
 
-            // Delete the backup file
-            if (File.Exists(tempManifestFile))
-                File.Delete(tempManifestFile);
+            int size = (int)(writer.BaseStream.Position - startPos);
+            writer.Write(size);
         }
 
         private void Read() {
 
             string manifestFile = Config.ManifestFile(_baseFileName);
             if (!File.Exists(manifestFile)) {
-                Write(); // Save current defaults
                 return;
             }
 
             FileStream fs = new FileStream(manifestFile, FileMode.Open, FileAccess.Read, FileShare.None, 1024, false);
             BinaryReader reader = new BinaryReader(fs);
+
+            // Get the size of the last manifest block
+            reader.BaseStream.Seek(-4, SeekOrigin.End);
+            int size = reader.ReadInt32();
+
+            // Now seek to that position and read it
+            reader.BaseStream.Seek(-size - 4, SeekOrigin.End);
+
             try {
+                _manifestVersion = reader.ReadInt32();
                 int num_versions = reader.ReadInt32();
                 for (int i = 0; i < num_versions; i++) {
                     _versions[i] = reader.ReadInt32();
